ingestion/requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
pydantic
kafka-python
python-multipart